backups/quantum_prompts_backup_20250401_083729/CRONOS/backups/system_backup_20250330_095733/requirements.txt
---
metadata:
  api_endpoints: []
  author: EVA & GUARANI
  backup_required: true
  category: requirements.txt
  changelog: []
  dependencies:
  - QUANTUM_PROMPTS
  - BIOS-Q
  description: Component of the EVA & GUARANI Quantum Unified System
  documentation_quality: 0.95
  encoding: utf-8
  ethical_validation: true
  last_updated: '2025-03-29'
  related_files: []
  required: true
  review_status: approved
  security_level: 0.95
  simulation_capable: false
  status: active
  subsystem: MASTER
  test_coverage: 0.9
  translation_status: completed
  type: text
  version: '8.0'
  windows_compatibility: true
---
# EVA & GUARANI Core Dependencies
pyyaml>=6.0.1
tqdm>=4.66.1
colorama>=0.4.6
rich>=13.7.0
# langdetect>=1.0.9 # Not compatible with Python 3.9+
# langid>=1.1.6 # Not compatible with Python 3.9+
lingua-language-detector>=2.0.0  # Modern language detection library
py3langid>=0.3.0                # Python 3 compatible fork of langid
argparse>=1.4.0
requests>=2.31.0
python-dotenv>=1.0.0
# pyperclip>=1.8.2  # For clipboard operations (conversation exporter) - Optional

# Bot Dependencies
python-telegram-bot==13.15      # Telegram bot API version compatible with our code
cryptography>=41.0.4            # For secure transactions

# Knowledge Acquisition Libraries
beautifulsoup4>=4.12.2   # HTML/XML parsing for web scraping
arxiv>=1.4.7             # arXiv API access
scholarly>=1.7.0         # Google Scholar API (optional)
lxml>=4.9.3              # XML processing

# OpenAI integration
openai>=1.10.0

# Format-specific processing
markdown>=3.5.1

# Testing and Development
pytest>=7.4.0
pytest-cov>=4.1.0

# Note: Module-specific dependencies are listed in their respective requirements.txt files
# For example: modules/translator_dev/requirements.txt

# Additional dependencies
transformers>=4.36.0
torch>=2.1.0
# sentencepiece>=0.1.99 # Optional, may fail on some Windows systems
diskcache>=5.6.3

# Core requirements
langdetect>=1.0.9

# Optional translation engines
transformers>=4.20.0
torch>=1.10.0

# Optional: For OpenAI integration
openai>=1.0.0

# Development utilities
black>=22.0.0
flake8>=6.0.0

# ========================================================================
# EVA & GUARANI - CORE REQUIREMENTS
# ========================================================================

# Core Libraries
numpy>=1.22.0
pandas>=1.3.5
scipy>=1.7.3
scikit-learn>=1.0.2
matplotlib>=3.5.1
networkx>=2.6.3

# Web & API
requests>=2.27.1
beautifulsoup4>=4.10.0
fastapi>=0.75.0
uvicorn>=0.17.0
pydantic>=1.9.0
httpx>=0.22.0

# Data Processing
nltk>=3.7
spacy>=3.2.2
gensim>=4.1.2
joblib>=1.1.0

# Storage & Serialization
PyYAML>=6.0
python-dotenv>=0.19.2
toml>=0.10.2
json5>=0.9.5

# Development & Testing
pytest>=7.0.0
black>=22.1.0
pylint>=2.12.2
mypy>=0.931
pytest-cov>=3.0.0

# Utilities
tqdm>=4.62.3
colorama>=0.4.4
python-slugify>=6.1.1
typing-extensions>=4.0.1
python-dateutil>=2.8.2
schedule>=1.1.0

# Visualization
plotly>=5.5.0
seaborn>=0.11.2

# Language Processing
lingua-language-detector>=1.0.0  # Language detection

# Optional - will be safely ignored if not available
# python-louvain>=0.16; python_version < "3.12"
# community>=0.15; python_version < "3.12"
# pyperclip>=1.8.2  # For clipboard operations (conversation exporter)

# ========================================================================
# END OF REQUIREMENTS
# ========================================================================

# Perplexity API integration
requests>=2.28.1
python-dotenv>=0.21.0

# MCP Server dependencies
websockets>=12.0
requests>=2.31.0
aiohttp>=3.9.1
psutil>=5.9.0

# Core dependencies
pytest==8.0.0
pytest-cov==4.1.0
typing-extensions==4.9.0
python-dotenv==1.0.1
pyyaml==6.0.1
jsonschema==4.21.1

# Logging and monitoring
structlog==24.1.0
prometheus-client==0.19.0

# Development dependencies
black==24.1.1
isort==5.13.2
mypy==1.8.0
flake8==7.0.0
pylint==3.0.3

# Additional dependencies
pathlib>=1.0.1
typing-extensions>=4.0.0
python-dotenv>=0.19.0
colorama>=0.4.4
tqdm>=4.62.0
psutil>=5.8.0
requests>=2.26.0
PyYAML>=6.0.0
python-json-logger>=2.0.2
watchdog>=2.1.6

# Development & Testing
pytest>=6.2.5
pytest-cov>=2.12.1
black>=21.9b0
flake8>=3.9.2
mypy>=0.910

# Optional Components
# Uncomment as needed
#transformers>=4.11.3  # For TRANSLATOR
#torch>=1.9.0  # For TRANSLATOR
#networkx>=2.6.3  # For ATLAS
#numpy>=1.21.2  # For NEXUS
#pandas>=1.3.3  # For CRONOS

# EVA & GUARANI - Quantum Unified System Dependencies
# Version: 8.0
# Last Updated: 2025-03-29

# Core Dependencies
websockets==12.0
asyncio==3.4.3
PyYAML==6.0.1
typing-extensions==4.9.0
python-dateutil==2.8.2

# Logging and Monitoring
structlog==24.1.0
python-json-logger==2.0.7

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0

# Type Checking
mypy==1.8.0
types-PyYAML==6.0.12.12

# Code Quality
black==24.1.1
isort==5.13.2
flake8==7.0.0

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0
autodoc==0.5.0

# Security
cryptography==42.0.2
python-jose==3.3.0
passlib==1.7.4

# Windows Compatibility
pywin32==306; sys_platform == 'win32'
